Re #2407 Move ExtraDirs out of Stack.Types.Config
