+++ package.yaml
@@ -282,6 +282,7 @@ library:
   - Stack.Types.DumpPackage
   - Stack.Types.EnvConfig
   - Stack.Types.EnvSettings
+  - Stack.Types.ExtraDirs
   - Stack.Types.GHCDownloadInfo
   - Stack.Types.GHCVariant
   - Stack.Types.GhcOptionKey
@@ -296,6 +297,7 @@ library:
   - Stack.Types.PackageName
   - Stack.Types.Platform
   - Stack.Types.Project
+  - Stack.Types.ProjectAndConfigMonoid
   - Stack.Types.ProjectConfig
   - Stack.Types.PvpBounds
   - Stack.Types.Resolver
